refactor: extract analysisFolderCounts query into its own file
run.ts was 481 lines, exceeding the 400-line project limit. Move the
analysisFolderCounts query (types, refs, resolver) to analysis-folder-counts.ts.
run.ts drops to 278 lines. Auto-import picks up the new file automatically.
